_to parameter on staking functions + migration output optionality
specify _to recipient for staking functions, enabling them to be used as transfers

allow migrator to choose output token

+++ contracts/migration/OlympusTokenMigrator.sol
@@ -121,7 +121,11 @@ contract OlympusTokenMigrator is Ownable {
     }
 
     // migrate OHMv1, sOHMv1, or wsOHM for OHMv2, sOHMv2, or gOHM
-    function migrate(uint256 _amount, TYPE _from, TYPE _to) external {
+    function migrate(
+        uint256 _amount,
+        TYPE _from,
+        TYPE _to
+    ) external {
         uint256 sAmount = _amount;
         uint256 wAmount = oldwsOHM.sOHMTowOHM(_amount);
 
@@ -145,21 +149,21 @@ contract OlympusTokenMigrator is Ownable {
 
     // migrate all tokens held
     function migrateAll(TYPE _to) external {
-        uint ohmBal = oldOHM.balanceOf(msg.sender);
-        uint sOHMBal = oldsOHM.balanceOf(msg.sender);
-        uint wsOHMBal = oldwsOHM.balanceOf(msg.sender);
+        uint256 ohmBal = oldOHM.balanceOf(msg.sender);
+        uint256 sOHMBal = oldsOHM.balanceOf(msg.sender);
+        uint256 wsOHMBal = oldwsOHM.balanceOf(msg.sender);
 
-        if(ohmBal > 0) {
+        if (ohmBal > 0) {
             oldOHM.safeTransferFrom(msg.sender, address(this), ohmBal);
         }
-        if(sOHMBal > 0) {
+        if (sOHMBal > 0) {
             oldsOHM.safeTransferFrom(msg.sender, address(this), sOHMBal);
         }
-        if(wsOHMBal > 0) {
+        if (wsOHMBal > 0) {
             oldwsOHM.safeTransferFrom(msg.sender, address(this), wsOHMBal);
         }
 
-        uint wAmount = wsOHMBal.add( oldwsOHM.sOHMTowOHM( ohmBal.add(sOHMBal) ) );
+        uint256 wAmount = wsOHMBal.add(oldwsOHM.sOHMTowOHM(ohmBal.add(sOHMBal)));
         if (ohmMigrated) {
             require(oldSupply >= oldOHM.totalSupply(), "OHMv1 minted");
             _send(wAmount, _to);
@@ -169,8 +173,8 @@ contract OlympusTokenMigrator is Ownable {
     }
 
     // send preferred token
-    function _send(uint wAmount, TYPE _to) internal {
-        if(_to == TYPE.WRAPPED) {
+    function _send(uint256 wAmount, TYPE _to) internal {
+        if (_to == TYPE.WRAPPED) {
             gOHM.safeTransfer(msg.sender, wAmount);
         } else if (_to == TYPE.STAKED) {
             newStaking.unwrap(msg.sender, wAmount);
